Complete backlog section 8: Cross-Platform & Build Improvements
- Add CMake as alternative build system (CMakeLists.txt + libft/CMakeLists.txt)
  with native Debug/Release build types and target-scoped strict warnings
- Fix all compiler warnings with stricter flags: -Wpedantic, -Wshadow,
  -Wconversion added to CFLAGS with -Werror (38 warnings fixed across 9 files)
- Add `make debug` target with ASan+UBSan (-fsanitize=address,undefined, -O0)
- Add `make release` target with -O3 -DNDEBUG (no debug symbols)
- Add multi-distro CI matrix: Ubuntu 22.04, Ubuntu 24.04, Debian 12, Fedora 39
- Document SDL2 migration path in docs/SDL2_MIGRATION.md (5 files affected,
  full API mapping provided)

+++ src/ft_cylinder.c
@@ -193,14 +193,14 @@ void ft_intersect_ray_cylinder(t_shape *sh, t_ray *ray)
 
 	ft_cylinder_calc_one(sh, ray, &tmp, &dot);
 	ft_cylinder_calc_two(&calc, ray, &tmp, &dot);
-	if ((ret = ft_cylinder_calc_three(&calc, &dist, &dot, ray)) != 0)
+	if ((ret = (int)ft_cylinder_calc_three(&calc, &dist, &dot, ray)) != 0)
 	{
 		ray->lenght = -1;
 		best_t = INFINITY;
 		ft_cylinder_cap_check(sh, ray, &best_t);
 		return;
 	}
-	if ((ret = ft_cylinder_calc_five(sh, &dist, &dot, ray)) != 0)
+	if ((ret = (int)ft_cylinder_calc_five(sh, &dist, &dot, ray)) != 0)
 	{
 		best_t = ray->lenght;
 		ft_cylinder_cap_check(sh, ray, &best_t);
